+++ kenya-trainings-menu/options.js
@@ -3,21 +3,26 @@ module.exports = function () {
         'tree_transplanting': {
             'en-ke': 'Tree Transplanting\n',
             'sw': 'Kupanda Miti\n',
-            enabled: true
+            enabled: false
         },
         'tree_bag_planting': {
             'en-ke': 'Tree Bag Planting\n',
             'sw': 'Kupanda miti mifukoni\n',
-            enabled: true
+            enabled: false
         },
         'tree_socketing': {
             'en-ke': 'Tree Socketing\n',
             'sw': 'Socketing Miti\n',
-            enabled: true
+            enabled: false
         },
         'sorghum_weeding': {
             'en-ke': 'Sorghum Weeding\n',
             'sw': 'Kupalilia Wimbi\n',
+            enabled: false
+        },
+        'maize_intercorp': {
+            'en-ke': 'Maize Intercrop\n',
+            'sw': 'Kupanda Mahindi/Maharagwe\n',
             enabled: true
         },
         'maize_topdress': {
@@ -25,9 +30,9 @@ module.exports = function () {
             'sw': 'TopDress\n',
             enabled: true
         },
-        'maize_intercorp': {
-            'en-ke': 'Maize Intercrop\n',
-            'sw': 'Kupanda Mahindi/Maharagwe\n',
+        'pest_mitigation': {
+            'en-ke': 'Pest Mitigation\n',
+            'sw': 'Wadudu/Magonjwa\n',
             enabled: true
         },
         'maize_harvest': {
@@ -35,30 +40,25 @@ module.exports = function () {
             'sw': 'Kuvuna Mahindi\n',
             enabled: true
         },
-        'pest_mitigation': {
-            'en-ke': 'Pest Mitigation\n',
-            'sw': 'Wadudu/Magonjwa\n',
-            enabled: true
-        },
         'vegetables': {
             'en-ke': 'Vegetables\n',
             'sw': 'Kupanda Mboga\n',
-            enabled: true
+            enabled: false
         },
         'tatu_hadi_tatu': {
             'en-ke': 'Tatu Hadi Tatu\n',
             'sw': 'Tatu Hadi Tatu\n',
-            enabled: true
+            enabled: false
         },
         'soil_training': {
             'en-ke': 'Soil Fertility\n',
             'sw': 'Udongo bora ulio na afya na rotuba\n',
-            enabled: true
+            enabled: false
         },
         'nutrition_training': {
             'en-ke': 'Dietary Diversity (Nutrition)\n',
             'sw': 'Lishe Bora\n',
-            enabled: true
+            enabled: false
         },
     };
 };