+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 All notable changes to this project will be documented in this file. 
 
 ## New Version
-
+## v1.10.6
+* [ESWFA-113] (https://oneacrefund.atlassian.net/browse/ESWFA-113) Fix the Kenya Impact Trainings
+## v1.10.5
+* [MOB-407] (https://oneacrefund.atlassian.net/browse/THD-5670) Switching on of Training service on USSD & 2-Way SMS
+## v1.10.4
+* [MOB-407] (https://oneacrefund.atlassian.net/browse/MOB-407) Handle additional Validation for wrong shs
 ## v1.10.3
 * [ESWFA-93](https://oneacrefund.atlassian.net/browse/ESWFA-93) Add an IPP option to the non farmer USSD menu
 ## v1.10.2